max, add obj written

+++ Aggregate/EdgeObj.h
@@ -1,3 +1,5 @@
+#include <stdlib.h>
+#include <stddef.h>
 #ifndef EDGEOBJ_H_INCLUDED
 #define EDGEOBJ_H_INCLUDED
 
@@ -11,16 +13,16 @@ class EdgeObj {//object on edge, landmark and moving obj
 
 public:
 //location information:
-    int id;
-    int snid;         //start node id
-    int enid;         //end node id
-    int eid;          //edge id
+    size_t id;
+    size_t snid;         //start node id
+    size_t enid;         //end node id
+    size_t eid;          //edge id
     double dist2sn;   //distance to start node
     double dist2en;   //distance to start node
 
 
     EdgeObj(){
-        this->id = -1;
+        this->id = (size_t) -1;
     }
 
     bool operator < (const EdgeObj& objItem) const{
